refactored the code design
