+++ app/src/main/java/com/amrdeveloper/linkhub/data/parser/ImportExportFileParser.kt
@@ -7,8 +7,9 @@ import com.amrdeveloper.linkhub.data.source.LinkRepository
 import com.amrdeveloper.linkhub.util.UiPreferences
 
 interface ImportExportFileParser {
-    object ImportExportFileParserFactory {
-        fun getInstance(fileType: ImportExportFileType): ImportExportFileParser {
+
+    companion object {
+        fun getDataParser(fileType: ImportExportFileType): ImportExportFileParser {
             return when (fileType) {
                 ImportExportFileType.JSON -> JsonImportExportFileParser()
                 ImportExportFileType.HTML -> HtmlImportExportFileParser()
